Training list changes (#63)
* Layout changes

* App is not crashing after deleting training

* fix
